[Feature] Add Support for item textures higher than 65,535 (#3494)
* [Feature] Add Support for item textures higher than 65,535

# Notes
- We were previously using `uint16` which has a max of `65,535`, but some of the new Live textures that are being used have values way beyond this.
- Updates quest API, commands, and database tables that also used `uint16`.

* Update version.h

+++ zone/lua_mob.h
@@ -386,9 +386,9 @@ public:
 	void TarGlobal(const char *varname, const char *value, const char *duration, int npc_id, int char_id, int zone_id);
 	void DelGlobal(const char *varname);
 	void SetSlotTint(int material_slot, int red_tint, int green_tint, int blue_tint);
-	void WearChange(uint8 material_slot, uint16 texture);
-	void WearChange(uint8 material_slot, uint16 texture, uint32 color);
-	void WearChange(uint8 material_slot, uint16 texture, uint32 color, uint32 heros_forge_model);
+	void WearChange(uint8 material_slot, uint32 texture);
+	void WearChange(uint8 material_slot, uint32 texture, uint32 color);
+	void WearChange(uint8 material_slot, uint32 texture, uint32 color, uint32 heros_forge_model);
 	void DoKnockback(Lua_Mob caster, uint32 push_back, uint32 push_up);
 	void AddNimbusEffect(int effect_id);
 	void RemoveNimbusEffect(int effect_id);
