+++ zone/npc.h
@@ -350,7 +350,7 @@ public:
 	int					GetClosestWaypoint(const glm::vec3& location);
 
 	uint32				GetEquippedItemFromTextureSlot(uint8 material_slot) const;	// returns item id
-	int32				GetEquipmentMaterial(uint8 material_slot) const;
+	uint32				GetEquipmentMaterial(uint8 material_slot) const;
 
 	void				NextGuardPosition();
 	void				SaveGuardSpot(bool ClearGuardSpot = false);
@@ -452,7 +452,7 @@ public:
 
 	bool GetDepop() { return p_depop; }
 
-	void NPCSlotTexture(uint8 slot, uint16 texture);	// Sets new material values for slots
+	void NPCSlotTexture(uint8 slot, uint32 texture);	// Sets new material values for slots
 
 	uint32 GetAdventureTemplate() const { return adventure_template_id; }
 	void AddSpellToNPCList(int16 iPriority, uint16 iSpellID, uint32 iType, int16 iManaCost, int32 iRecastDelay, int16 iResistAdjust, int8 min_hp, int8 max_hp);
@@ -648,9 +648,9 @@ protected:
 	uint32	equipment[EQ::invslot::EQUIPMENT_COUNT];	//this is an array of item IDs
 
 	uint32	herosforgemodel;			//this is the Hero Forge Armor Model (i.e 63 or 84 or 203)
-	uint16	d_melee_texture1;
+	uint32	d_melee_texture1;
 	//this is an item Material value
-	uint16	d_melee_texture2;			//this is an item Material value (offhand)
+	uint32	d_melee_texture2;			//this is an item Material value (offhand)
 	const char*	ammo_idfile;			//this determines projectile graphic "IT###" (see item field 'idfile')
 	uint8	prim_melee_type;			//Sets the Primary Weapon attack message and animation
 	uint8	sec_melee_type;				//Sets the Secondary Weapon attack message and animation
